fix:booth error log올라간것 지우기

+++ src/main/java/com/hiufestainfo/domain/booth/service/BoothService.java
@@ -1,15 +1,20 @@
 package com.hiufestainfo.domain.booth.service;
 
 import com.hiufestainfo.domain.booth.dto.BoothRequestDto;
+import com.hiufestainfo.domain.booth.dto.BoothResponseDto;
 import com.hiufestainfo.domain.booth.entity.Booth;
 import com.hiufestainfo.domain.booth.exception.BoothNotFoundException;
 import com.hiufestainfo.domain.booth.exception.CreateBoothBadRequestException;
 import com.hiufestainfo.domain.booth.repository.BoothRepository;
+import com.hiufestainfo.domain.user.entity.Role;
+import com.hiufestainfo.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -17,8 +22,14 @@ import java.util.List;
 public class BoothService {
     private final BoothRepository boothRepository;
     public String deleteBooth(Long boothId) {
-        boothRepository.deleteById(boothId);
-        return "Delete Success";
+        Optional<Booth> boothOptional = boothRepository.findById(boothId);
+
+        if (boothOptional.isPresent()) {
+            boothRepository.deleteById(boothId);
+            return "Delete Success";
+        } else {
+            throw new BoothNotFoundException();
+        }
     }
 
     public Booth createBooth(BoothRequestDto requestDto) {
@@ -26,7 +37,7 @@ public class BoothService {
         String boothNum = requestDto.getBoothNum();
         String intro = requestDto.getIntro();
         String boothName = requestDto.getBoothName();
-        String imageUrl= requestDto.getImageUrl();
+        String host = requestDto.getHost();
 
 
         // Booth 엔티티 생성
@@ -34,7 +45,7 @@ public class BoothService {
                 .boothNum(boothNum)
                 .boothName(boothName)
                 .intro(intro)
-                .imageUrl(imageUrl)
+                .host(host)
                 .build();
 
         // 생성된 Pub 엔티티를 저장하고 반환
@@ -44,12 +55,12 @@ public class BoothService {
 
     public Booth updateBooth(Long boothId, BoothRequestDto requestDto){
         Booth booth = boothRepository.findById(boothId)
-                .orElseThrow(() -> new IllegalArgumentException("Invalid pub ID: " + boothId));
+                .orElseThrow(() -> new BoothNotFoundException());
 
         String intro= requestDto.getIntro();
         String boothName = requestDto.getBoothName();
         String boothNum = requestDto.getBoothNum();
-        String imageUrl = requestDto.getImageUrl();
+        String host = requestDto.getHost();
 
         // 요청된 데이터로 엔티티를 수정합니다.
 
@@ -64,23 +75,34 @@ public class BoothService {
         if(boothName==null||(boothName=="")){
             boothName = booth.getBoothName();
         }
-        if(imageUrl==null||(imageUrl=="")){
-            imageUrl=booth.getImageUrl();
+        if(host==null||(host=="")){
+           host=booth.getHost();
         }
 
-        booth.updateBooth(boothNum,boothName,intro,imageUrl);
+        booth.updateBooth(boothNum,boothName,intro,host);
 
         // 수정된 엔티티를 저장하고 반환합니다.
         return boothRepository.save(booth);
     }
 
 
-    public List<Booth> getAllBooths (){
-        List<Booth> booths = boothRepository.findAll();
-        if (booths == null || booths.isEmpty()) {
-            throw new BoothNotFoundException();
+    public BoothResponseDto getAllBooths (User user){
+
+        Boolean isAdmin = false;
+        Role accountStatus = user.getAuthInfo().getRole();
+        if( accountStatus.getValue() == "GUEST"){
+            isAdmin = false;
+        } else if(accountStatus.getValue() == "ADMIN"){
+            isAdmin = true;
         }
-        return booths;
+
+        List<Booth> booths = boothRepository.findAll();
+        // booths가 null이면 빈 리스트로 초기화
+        if (booths == null) {
+            booths = new ArrayList<>();
+        }
+
+        return new BoothResponseDto(isAdmin, booths);
     }
 
 }
